requirements.txt
torch>=2.1.0
open-clip-torch>=2.24.0
timm>=0.9.0
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
tqdm>=4.65.0
wandb>=0.15.0
fvcore>=0.1.5
huggingface-hub>=0.20.0
